chore(app): remove deprecated `bin/www` (#418)
* refactor(bin/www): remove bin/www and inline into bootstrap/index

* refactor(server): update to use new bootstrap function

* chore(package.json): update commands
