+++ package.json
@@ -4,9 +4,9 @@
   "private": true,
   "scripts": {
     "build": "tsc -p tsconfig.build.json",
-    "start": "node -r ts-node/register/transpile-only -r tsconfig-paths/register build/bin/www.js",
+    "start": "node -r ts-node/register/transpile-only -r tsconfig-paths/register build/server.js",
     "dev:services": "docker compose up -d",
-    "dev": "source .env && ts-node-dev --respawn ./bin/www.js",
+    "dev": "source .env && ts-node-dev --respawn server.js",
     "test": "source .env.test && jest",
     "release": "npm version $npm_config_isomer_update && git push --tags",
     "lint": "npx eslint .",
